Update to Zephyr 4.2 (#333)

+++ lib/f_core/device/sensor/c_magnetometer.cpp
@@ -23,7 +23,7 @@ sensor_value CMagnetometer::GetSensorValue(sensor_channel chan) const {
             return magData.z;
         default:
             // Assert here since this should never occur
-            LOG_ERR("Invalid sensor channel (%d) called for magnetometer", chan);
+            LOG_ERR("Invalid sensor channel (%d) called for magnetometer", static_cast<int>(chan));
             k_oops();
             return {INT32_MIN, INT32_MIN};
     }
